表示状態を切り替えた際の拡縮を登録 withSwitchGroup関数を登録
