+++ src/scene_composer_extra/src/controllers/TagController.tsx
@@ -3,28 +3,12 @@ import {
   KnownComponentType,
 } from "@iot-app-kit/scene-composer";
 import { ISceneNodeInternal } from "@iot-app-kit/scene-composer/dist/src/store";
-import { MMDModelWrapper } from "../objects/mmd/MMDModelWrapper";
 import { SearchTagsCallback } from "../types/DataType";
-import { MeshUiButtonWrapper } from "../objects/three-mesh-ui/MeshUiButtonWrapper";
-import { MeshUiTextWrapper } from "../objects/three-mesh-ui/MeshUiTextWrapper";
 import { Object3D, Event, Scene } from "three/src/Three";
-import { GLTFModelWrapper } from "../objects/model/GLTFModelWrapper";
-import { HTMLModelWrapper } from "../objects/html/HTMLModelWrapper";
-import { RechartsModelWrapper } from "../objects/graph/RechartsModelWrapper";
-import {
-  Bar,
-  BarChart,
-  CartesianGrid,
-  Line,
-  LineChart,
-  XAxis,
-  YAxis,
-} from "recharts";
-import React from "react";
-import { ImageModelWrapper } from "../objects/model/ImageModelWrapper";
-import { TextureAtrasVideoWrapper } from "../objects/model/TextureAtrasVideoWrapper";
-import { GroupWrapper } from "../objects/group/GroupWrapper";
-import { ExtraObjectWrapperParameter } from "../objects/ExtraObjectWrapper";
+import { ReplaceTagBase } from "./tags/ReplaceTagBase";
+import { ReplaceTagBasic } from "./tags/ReplaceTagBasic";
+import { ReplaceTagGraph } from "./tags/ReplaceTagGraph";
+import { ReplaceTagUtility } from "./tags/ReplaceTagUtility";
 
 /**
  * タグを検索する
@@ -109,201 +93,7 @@ export class ReplaceContext {
   }
 }
 
-export class ReplaceTag {
-  // カメラ、レンダリング情報を持ったシーンオブジェクト
-  _rootScene: Scene;
-  // TwinMakerで管理されるタグの表示情報、アンカー情報
-  _anchor: IAnchorComponent;
-  // オプション: TwinMakerタグの3Dオブジェクト情報
-  _tag?: Object3D<Event>;
-  // オプション: 置き換え後のオブジェクトを配置するグループ
-  _parentGroup?: Object3D;
-
-  /**
-   * @param ref タグオブジェクトのRef ID
-   * @param anchor タグオブジェクトのアンカー情報(位置、色、アイコン情報などが入っている)
-   */
-  constructor(
-    rootScene: Scene,
-    anchor: IAnchorComponent,
-    tag?: Object3D<Event>,
-    parentGroup?: Object3D
-  ) {
-    this._rootScene = rootScene;
-    this._anchor = anchor;
-    this._tag = tag;
-    this._parentGroup = parentGroup;
-  }
-
-  /* タグからの入れ替え先に受け渡すパラメータを設定する */
-  parameter(tag: Object3D<Event>): ExtraObjectWrapperParameter {
-    return {
-      rootScene: this._rootScene,
-      position: tag.position,
-      rotate: tag.rotation,
-      scale: tag.scale,
-      anchor: this._anchor,
-      parentObject: this._parentGroup,
-    };
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトをグループに置き換える
-   */
-  get toGroup() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new GroupWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトをMMDに置き換える
-   *
-   * @param ref タグオブジェクトのRef ID
-   * @param anchor タグオブジェクトのアンカー情報(位置、色、アイコン情報などが入っている)
-   * @returns MMDの管理クラス
-   */
-  get toMMD() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new MMDModelWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトをGLTF, GLBに置き換える
-   *
-   * @returns GLTF, GLBの管理クラス
-   */
-  get toGLTF() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new GLTFModelWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトをボタンに置き換える
-   *
-   * @returns MMDの管理クラス
-   */
-  get toButton() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new MeshUiButtonWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトをテキストに置き換える
-   *
-   * @returns MMDの管理クラス
-   */
-  get toText() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new MeshUiTextWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトをHTMLに置き換える
-   */
-  get toHTML() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new HTMLModelWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトを2D画像に置き換える
-   */
-  get toImage() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new ImageModelWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * テクスチャアトラスを使ったビデオに置き換える
-   */
-  get toTextureAtrasVideo() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new TextureAtrasVideoWrapper(this.parameter(tag));
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトを線グラフに置き換える
-   */
-  get toLineChart() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new RechartsModelWrapper(
-        this.parameter(tag)
-      ).updateRachartsComponent(
-        LineChart,
-        <>
-          <XAxis dataKey="x" />
-          <YAxis />
-          <CartesianGrid stroke="#eee" strokeDasharray="5 5" />
-          <Line
-            type="monotone"
-            dataKey="y"
-            stroke="#8884d8"
-            isAnimationActive={false}
-          />
-        </>
-      );
-    }
-    return undefined;
-  }
-
-  /**
-   * TwinMakerのタグオブジェクトを棒グラフに置き換える
-   */
-  get toBarChart() {
-    const tag = this._tag;
-    if (tag) {
-      tag.visible = false;
-      return new RechartsModelWrapper(
-        this.parameter(tag)
-      ).updateRachartsComponent(
-        BarChart,
-        <>
-          <CartesianGrid stroke="#eee" strokeDasharray="5 5" />
-          <XAxis dataKey="x" />
-          <YAxis />
-          <Bar
-            type="monotone"
-            dataKey="y"
-            fill="#8884d8"
-            isAnimationActive={false}
-          />
-        </>
-      );
-    }
-    return undefined;
-  }
-}
+/** タグの変換関数の一覧を登録する */
+export class ReplaceTag extends ReplaceTagUtility(
+  ReplaceTagGraph(ReplaceTagBasic(ReplaceTagBase))
+) {}
